requirements.txt
nest_asyncio
streamlit
langchain
langchain_mcp_adapters
langgraph
langchain_google_genai
google-generativeai 
python-dotenv